Distribute preset factory methods into components
