+++ src/main/java/de/vinado/wicket/participate/ui/event/ParticipantColumnListDecoratorFactory.java
@@ -1,0 +1,10 @@
+package de.vinado.wicket.participate.ui.event;
+
+/**
+ * @author Vincent Nadoll
+ */
+@FunctionalInterface
+public interface ParticipantColumnListDecoratorFactory {
+
+    ParticipantColumnList decorate(ParticipantColumnList preset);
+}
